+++ device-bootloaders/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [SW-18460] Add more error logs for debugging.
 - [SW-15352] Check PMIC board type in FOTA process.
+- [SW-18602] Added support to rescan a partition after a FW image write.
 ### Changed
 - [SW-18208] Enabling PMIC Firmware update.
 ### Deprecated